moved loss stuff out of workhorse
